Refactor structure of modules.
