+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "appchain-registry-wrapper"
-version = "2.1.0"
+version = "3.0.0"
 authors = ["Octopus Network"]
 edition = "2021"
