+++ appchain-registry/src/user_actions/sudo_actions.rs
@@ -1,5 +1,20 @@
-use crate::{interfaces::SudoActions, *};
+use crate::*;
 use std::{convert::TryFrom, str::FromStr};
+
+pub trait SudoActions {
+    /// Set public key of owner.
+    fn set_owner_pk(&mut self, public_key: String);
+    /// Create subaccount for a specific appchain.
+    fn create_anchor_account(&mut self, appchain_id: AppchainId);
+    /// Force change state of an appchain.
+    fn force_change_appchain_state(&mut self, appchain_id: AppchainId, state: AppchainState);
+    /// Pause asset transfer in this contract.
+    fn pause_asset_transfer(&mut self);
+    /// Resume asset transfer in this contract.
+    fn resume_asset_transfer(&mut self);
+    /// Force remove an appchain.
+    fn force_remove_appchain(&mut self, appchain_id: AppchainId);
+}
 
 #[near_bindgen]
 impl SudoActions for AppchainRegistry {
